refactor authorization for sanity
