+++ pkg/authorization/authorizer/attributes_builder.go
@@ -1,0 +1,53 @@
+package authorizer
+
+import (
+	"net/http"
+	"strings"
+
+	kapi "github.com/GoogleCloudPlatform/kubernetes/pkg/api"
+	kapiserver "github.com/GoogleCloudPlatform/kubernetes/pkg/apiserver"
+)
+
+type openshiftAuthorizationAttributeBuilder struct {
+	contextMapper kapi.RequestContextMapper
+	infoResolver  *kapiserver.APIRequestInfoResolver
+}
+
+func NewAuthorizationAttributeBuilder(contextMapper kapi.RequestContextMapper, infoResolver *kapiserver.APIRequestInfoResolver) AuthorizationAttributeBuilder {
+	return &openshiftAuthorizationAttributeBuilder{contextMapper, infoResolver}
+}
+
+func (a *openshiftAuthorizationAttributeBuilder) GetAttributes(req *http.Request) (AuthorizationAttributes, error) {
+	// any url that starts with an API prefix and is more than one step long is considered to be a resource URL.
+	// That means that /api is non-resource, /api/v1beta1 is resource, /healthz is non-resource, and /swagger/anything is non-resource
+	urlSegments := splitPath(req.URL.Path)
+	isResourceURL := (len(urlSegments) > 1) && a.infoResolver.APIPrefixes.Has(urlSegments[0])
+
+	if !isResourceURL {
+		return DefaultAuthorizationAttributes{
+			Verb:           strings.ToLower(req.Method),
+			NonResourceURL: true,
+			URL:            req.URL.Path,
+		}, nil
+	}
+
+	requestInfo, err := a.infoResolver.GetAPIRequestInfo(req)
+	if err != nil {
+		return nil, err
+	}
+
+	// TODO reconsider special casing this.  Having the special case hereallow us to fully share the kube
+	// APIRequestInfoResolver without any modification or customization.
+	if (requestInfo.Resource == "projects") && (len(requestInfo.Name) > 0) {
+		requestInfo.Namespace = requestInfo.Name
+	}
+
+	return DefaultAuthorizationAttributes{
+		Verb:              requestInfo.Verb,
+		Resource:          requestInfo.Resource,
+		ResourceName:      requestInfo.Name,
+		RequestAttributes: nil,
+		NonResourceURL:    false,
+		URL:               req.URL.Path,
+	}, nil
+}
