+++ pkg/authorization/authorizer/bootstrap_policy.go
@@ -1,0 +1,233 @@
+package authorizer
+
+import (
+	kapi "github.com/GoogleCloudPlatform/kubernetes/pkg/api"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
+
+	authorizationapi "github.com/openshift/origin/pkg/authorization/api"
+)
+
+func GetBootstrapPolicy(masterNamespace string) *authorizationapi.Policy {
+	return &authorizationapi.Policy{
+		ObjectMeta: kapi.ObjectMeta{
+			Name:              authorizationapi.PolicyName,
+			Namespace:         masterNamespace,
+			CreationTimestamp: util.Now(),
+			UID:               util.NewUUID(),
+		},
+		LastModified: util.Now(),
+		Roles: map[string]authorizationapi.Role{
+			"cluster-admin": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "cluster-admin",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{
+						Verbs:     util.NewStringSet(authorizationapi.VerbAll),
+						Resources: util.NewStringSet(authorizationapi.ResourceAll),
+					},
+					{
+						Verbs:           util.NewStringSet(authorizationapi.VerbAll),
+						NonResourceURLs: util.NewStringSet(authorizationapi.NonResourceAll),
+					},
+				},
+			},
+			"admin": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "admin",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{
+						Verbs:     util.NewStringSet("get", "list", "watch", "create", "update", "delete"),
+						Resources: util.NewStringSet(authorizationapi.OpenshiftExposedGroupName, authorizationapi.PermissionGrantingGroupName, authorizationapi.KubeExposedGroupName),
+					},
+					{
+						Verbs:     util.NewStringSet("get", "list", "watch"),
+						Resources: util.NewStringSet(authorizationapi.PolicyOwnerGroupName, authorizationapi.KubeAllGroupName),
+					},
+				},
+			},
+			"edit": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "edit",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{
+						Verbs:     util.NewStringSet("get", "list", "watch", "create", "update", "delete"),
+						Resources: util.NewStringSet(authorizationapi.OpenshiftExposedGroupName, authorizationapi.KubeExposedGroupName),
+					},
+					{
+						Verbs:     util.NewStringSet("get", "list", "watch"),
+						Resources: util.NewStringSet(authorizationapi.KubeAllGroupName),
+					},
+				},
+			},
+			"view": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "view",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{
+						Verbs:     util.NewStringSet("get", "list", "watch"),
+						Resources: util.NewStringSet(authorizationapi.OpenshiftExposedGroupName, authorizationapi.KubeAllGroupName),
+					},
+				},
+			},
+			"basic-user": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "view-self",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{Verbs: util.NewStringSet("get"), Resources: util.NewStringSet("users"), ResourceNames: util.NewStringSet("~")},
+					{Verbs: util.NewStringSet("list"), Resources: util.NewStringSet("projects")},
+				},
+			},
+			"cluster-status": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "cluster-status",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{
+						Verbs:           util.NewStringSet("get"),
+						NonResourceURLs: util.NewStringSet("/healthz", "/version", "/api", "/osapi"),
+					},
+				},
+			},
+			"system:deployer": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "system:deployer",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{
+						Verbs:     util.NewStringSet(authorizationapi.VerbAll),
+						Resources: util.NewStringSet(authorizationapi.ResourceAll),
+					},
+				},
+			},
+			"system:component": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "system:component",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{
+						Verbs:     util.NewStringSet(authorizationapi.VerbAll),
+						Resources: util.NewStringSet(authorizationapi.ResourceAll),
+					},
+				},
+			},
+			"system:delete-tokens": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "system:delete-tokens",
+					Namespace: masterNamespace,
+				},
+				Rules: []authorizationapi.PolicyRule{
+					{
+						Verbs:     util.NewStringSet("delete"),
+						Resources: util.NewStringSet("oauthaccesstoken", "oauthauthorizetoken"),
+					},
+				},
+			},
+		},
+	}
+}
+
+func GetBootstrapPolicyBinding(masterNamespace string) *authorizationapi.PolicyBinding {
+	return &authorizationapi.PolicyBinding{
+		ObjectMeta: kapi.ObjectMeta{
+			Name:              masterNamespace,
+			Namespace:         masterNamespace,
+			CreationTimestamp: util.Now(),
+			UID:               util.NewUUID(),
+		},
+		LastModified: util.Now(),
+		PolicyRef:    kapi.ObjectReference{Namespace: masterNamespace},
+		RoleBindings: map[string]authorizationapi.RoleBinding{
+			"system:component-binding": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "system:component-binding",
+					Namespace: masterNamespace,
+				},
+				RoleRef: kapi.ObjectReference{
+					Name:      "system:component",
+					Namespace: masterNamespace,
+				},
+				Users: util.NewStringSet("system:openshift-client", "system:kube-client"),
+			},
+			"system:deployer-binding": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "system:deployer-binding",
+					Namespace: masterNamespace,
+				},
+				RoleRef: kapi.ObjectReference{
+					Name:      "system:deployer",
+					Namespace: masterNamespace,
+				},
+				Users: util.NewStringSet("system:openshift-deployer"),
+			},
+			"cluster-admin-binding": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "cluster-admin-binding",
+					Namespace: masterNamespace,
+				},
+				RoleRef: kapi.ObjectReference{
+					Name:      "cluster-admin",
+					Namespace: masterNamespace,
+				},
+				Users: util.NewStringSet("system:admin"),
+			},
+			"basic-user-binding": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "basic-user-binding",
+					Namespace: masterNamespace,
+				},
+				RoleRef: kapi.ObjectReference{
+					Name:      "basic-user",
+					Namespace: masterNamespace,
+				},
+				Groups: util.NewStringSet("system:authenticated"),
+			},
+			"insecure-cluster-admin-binding": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "insecure-cluster-admin-binding",
+					Namespace: masterNamespace,
+				},
+				RoleRef: kapi.ObjectReference{
+					Name:      "cluster-admin",
+					Namespace: masterNamespace,
+				},
+				// TODO until we decide to enforce policy, simply allow every one access
+				Groups: util.NewStringSet("system:authenticated", "system:unauthenticated"),
+			},
+			"system:delete-tokens-binding": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "system:delete-tokens-binding",
+					Namespace: masterNamespace,
+				},
+				RoleRef: kapi.ObjectReference{
+					Name:      "system:delete-tokens",
+					Namespace: masterNamespace,
+				},
+				Groups: util.NewStringSet("system:authenticated", "system:unauthenticated"),
+			},
+			"cluster-status-binding": {
+				ObjectMeta: kapi.ObjectMeta{
+					Name:      "cluster-status-binding",
+					Namespace: masterNamespace,
+				},
+				RoleRef: kapi.ObjectReference{
+					Name:      "cluster-status",
+					Namespace: masterNamespace,
+				},
+				Groups: util.NewStringSet("system:authenticated", "system:unauthenticated"),
+			},
+		},
+	}
+}
